+++ server/src/exception/http-exception.filter.ts
